About halfway to using cached matrices for collision detection. Implemented on client side, and cleaned up server side quite a bit. No more link_t garbage; use GList to manage entities in sectors (like Q3) -- avoiding confusing reuse of term 'area'.

+++ src/client/cl_predict.c
@@ -78,7 +78,7 @@ int32_t Cl_PointContents(const vec3_t point) {
 		// FIXME: To be entirely correct, we should use cl.lerp here and
 		// interpolate the origin and angles.
 
-		const int32_t num = (cl.frame.entity_state + i) & ENTITY_STATE_MASK;
+		const uint32_t num = (cl.frame.entity_state + i) & ENTITY_STATE_MASK;
 		const entity_state_t *ent = &cl.entity_states[num];
 
 		if (ent->solid < SOLID_BOX) // only clip to boxes, missiles and BSP models
@@ -111,7 +111,7 @@ static void Cl_ClipTraceToEntities(cl_trace_t *trace) {
 
 	for (uint16_t i = 0; i < cl.frame.num_entities; i++) {
 
-		const uint16_t num = (cl.frame.entity_state + i) & ENTITY_STATE_MASK;
+		const uint32_t num = (cl.frame.entity_state + i) & ENTITY_STATE_MASK;
 		const entity_state_t *ent = &cl.entity_states[num];
 
 		// FIXME: To be entirely correct, we should use cl.lerp here and
